+++ changelog.md
@@ -1,3 +1,7 @@
+1.0.3
+* If you visit a Soul Healer and you have no Soul Sickness and no XP to recover, the PvE death prompt (which tells you to visit Soul Healer) will now disappear, and this app will continue to function.
+* Added display of player's character level (if entered)
+
 1.0.2
 * Fixed XP chart for level 57-59
 * If starting AP was not entered during startup, the player will no longer be prompted to manually input AP, during soul healing or manual EXP update, if player had been attacked by another player